Address code reviews
